+++ packages/services/src/view-service/index.ts
@@ -30,6 +30,7 @@ import { walletId } from './wallet-id.js';
 import { witness } from './witness.js';
 import { witnessAndBuild } from './witness-and-build.js';
 import { transparentAddress } from './transparent-address.js';
+import { latestSwaps } from './latest-swaps.js';
 
 export type Impl = ServiceImpl<typeof ViewService>;
 
@@ -64,4 +65,5 @@ export const viewImpl: Impl = {
   witness,
   witnessAndBuild,
   transparentAddress,
+  latestSwaps,
 };