+++ packages/storage/src/indexed-db/index.ts
@@ -517,9 +517,24 @@ export class IndexedDb implements IndexedDbInterface {
 
   async getAltGasPrices(): Promise<GasPrices[]> {
     const allGasPrices = await this.db.getAll('GAS_PRICES');
+    const usdcPriorityAssetId = 'drPksQaBNYwSOzgfkGOEdrd4kEDkeALeh58Ps+7cjQs=';
+
+    // Retrieve gas prices from the database and prioritize USDC as the preferred asset.
     return allGasPrices
       .map(gp => GasPrices.fromJson(gp))
-      .filter(gp => !gp.assetId?.equals(this.stakingTokenAssetId));
+      .filter(gp => gp.assetId?.inner && !gp.assetId.equals(this.stakingTokenAssetId))
+      .sort((a, b) => {
+        const assetA = a.assetId?.inner ? uint8ArrayToBase64(a.assetId.inner) : '';
+        const assetB = b.assetId?.inner ? uint8ArrayToBase64(b.assetId.inner) : '';
+
+        if (assetA === usdcPriorityAssetId) {
+          return -1;
+        }
+        if (assetB === usdcPriorityAssetId) {
+          return 1;
+        }
+        return 0;
+      });
   }
 
   async saveGasPrices(value: Required<PlainMessage<GasPrices>>): Promise<void> {
@@ -600,6 +615,7 @@ export class IndexedDb implements IndexedDbInterface {
   async *getOwnedPositionIds(
     positionState: PositionState | undefined,
     tradingPair: TradingPair | undefined,
+    subaccount: AddressIndex | undefined,
   ) {
     yield* new ReadableStream({
       start: async cont => {
@@ -608,7 +624,10 @@ export class IndexedDb implements IndexedDbInterface {
           const position = Position.fromJson(cursor.value.position);
           if (
             (!positionState || positionState.equals(position.state)) &&
-            (!tradingPair || tradingPair.equals(position.phi?.pair))
+            (!tradingPair || tradingPair.equals(position.phi?.pair)) &&
+            (!subaccount ||
+              (cursor.value.subaccount &&
+                subaccount.equals(AddressIndex.fromJson(cursor.value.subaccount))))
           ) {
             cont.enqueue(PositionId.fromJson(cursor.value.id));
           }
@@ -619,16 +638,25 @@ export class IndexedDb implements IndexedDbInterface {
     });
   }
 
-  async addPosition(positionId: PositionId, position: Position): Promise<void> {
+  async addPosition(
+    positionId: PositionId,
+    position: Position,
+    subaccount?: AddressIndex,
+  ): Promise<void> {
     assertPositionId(positionId);
     const positionRecord = {
       id: positionId.toJson() as Jsonified<PositionId>,
       position: position.toJson() as Jsonified<Position>,
+      subaccount: subaccount && (subaccount.toJson() as Jsonified<AddressIndex>),
     };
     await this.u.update({ table: 'POSITIONS', value: positionRecord });
   }
 
-  async updatePosition(positionId: PositionId, newState: PositionState): Promise<void> {
+  async updatePosition(
+    positionId: PositionId,
+    newState: PositionState,
+    subaccount?: AddressIndex,
+  ): Promise<void> {
     assertPositionId(positionId);
     const key = uint8ArrayToBase64(positionId.inner);
     const positionRecord = await this.db.get('POSITIONS', key);
@@ -645,6 +673,7 @@ export class IndexedDb implements IndexedDbInterface {
       value: {
         id: positionId.toJson() as Jsonified<PositionId>,
         position: position.toJson() as Jsonified<Position>,
+        subaccount: subaccount ? (subaccount.toJson() as Jsonified<AddressIndex>) : undefined,
       },
     });
   }
@@ -1037,5 +1066,16 @@ export class IndexedDb implements IndexedDbInterface {
         },
         new Amount({ hi: 0n, lo: 0n }),
       );
+  }
+
+  async getPosition(positionId: PositionId): Promise<Position | undefined> {
+    assertPositionId(positionId);
+    const position = await this.db.get('POSITIONS', uint8ArrayToBase64(positionId.inner));
+
+    if (!position) {
+      return undefined;
+    }
+
+    return Position.fromJson(position.position);
   }
 }