+++ packages/wasm/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @penumbra-zone/wasm
 
+## 38.0.0
+
+### Patch Changes
+
+- Updated dependencies [6869c52]
+- Updated dependencies [29dd11a]
+  - @penumbra-zone/types@28.0.0
+  - @penumbra-zone/protobuf@7.1.0
+  - @penumbra-zone/bech32m@12.0.0
+
 ## 37.1.0
 
 ### Minor Changes